aI/requirements.txt
flask
google-generativeai
python-dotenv
requests
# pydantic (optional, for strong validation)
